Blood_supply/hospital_service/requirements.txt
Django==5.0.7
djangorestframework==3.15.1
kafka-python==2.0.2
psycopg2-binary==2.9.9
django-environ==0.11.2